+++ src/main/java/com/vanym/paniclecraft/client/ClientProxy.java
@@ -1,18 +1,27 @@
 package com.vanym.paniclecraft.client;
 
 import com.vanym.paniclecraft.Core;
+import com.vanym.paniclecraft.client.command.ClientCommandMod3;
+import com.vanym.paniclecraft.client.command.ClientCommandVersion;
+import com.vanym.paniclecraft.command.CommandMod3;
 import com.vanym.paniclecraft.core.CommonProxy;
 import com.vanym.paniclecraft.core.ModConfig;
 import com.vanym.paniclecraft.core.component.IModComponent;
 
 import cpw.mods.fml.relauncher.Side;
 import cpw.mods.fml.relauncher.SideOnly;
+import net.minecraftforge.client.ClientCommandHandler;
 
 @SideOnly(Side.CLIENT)
 public class ClientProxy extends CommonProxy {
     
+    public ClientCommandMod3 command;
+    
     @Override
     public void preInit(ModConfig config) {
+        this.command = new ClientCommandMod3();
+        this.command.addSubCommand(new ClientCommandVersion());
+        ClientCommandHandler.instance.registerCommand(this.command);
         for (IModComponent component : Core.instance.getComponents()) {
             component.preInitClient(config);
         }
@@ -33,5 +42,10 @@ public class ClientProxy extends CommonProxy {
         for (IModComponent component : Core.instance.getComponents()) {
             component.configChangedClient(config);
         }
+    }
+    
+    @Override
+    public CommandMod3 getCommand() {
+        return this.command;
     }
 }