+++ src/main/java/com/vanym/paniclecraft/client/command/ClientCommandMod3.java
@@ -1,0 +1,12 @@
+package com.vanym.paniclecraft.client.command;
+
+import com.vanym.paniclecraft.command.CommandMod3;
+
+public class ClientCommandMod3 extends CommandMod3 {
+    
+    public static final String NAME = CommandMod3.NAME + "client";
+    
+    public ClientCommandMod3() {
+        super(NAME);
+    }
+}