+++ src/main/java/com/vanym/paniclecraft/client/command/ClientCommandVersion.java
@@ -1,0 +1,13 @@
+package com.vanym.paniclecraft.client.command;
+
+import com.vanym.paniclecraft.command.CommandVersion;
+
+import net.minecraft.command.ICommandSender;
+
+public class ClientCommandVersion extends CommandVersion {
+    
+    @Override
+    public boolean canCommandSenderUseCommand(ICommandSender sender) {
+        return true;
+    }
+}