+++ src/main/java/com/vanym/paniclecraft/command/CommandMod3.java
@@ -8,8 +8,15 @@ public class CommandMod3 extends TreeCommandBase {
     
     public static final String NAME = DEF.MOD_ID;
     
+    protected final String name;
+    
     public CommandMod3() {
-        this.path = new String[]{this.getCommandName()};
+        this(NAME);
+    }
+    
+    protected CommandMod3(String name) {
+        this.name = name;
+        this.path = new String[]{name};
     }
     
     @Override
@@ -19,6 +26,6 @@ public class CommandMod3 extends TreeCommandBase {
     
     @Override
     public String getCommandName() {
-        return NAME;
+        return this.name;
     }
 }